fix: isNotEmpty able to handle non-string fields
If a non-string field like 'weight' were checked with this validation the result would be an error, since zod demans a string for this checking. So now non-string fields are only checked for emptiness if they were not informed on the request - when they receive the defaultValue "" and this way are able to be checked.