refactor: Log instance-hardcoded props
the properties to be shown by the log 'build_info', 'commitHash' and 'requestId' now get their value on class instantiation, but not depending on arguments. This was made for enabling the creation of log objects that allow 'stack' to continue being the last property on logging and for avoiding pollution on creating them on the middleware.